fix: レイアウトネストの調整
